refactor: separate component for repo add
